requirements.txt
paho-mqtt==1.6.1
aiohttp==3.9.5
python-mpv==1.0.8
PySide6==6.7.0
watchdog==4.0.1
python-dotenv==1.0.1
